requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
scikit-image==0.24.0
